src/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.7
certifi==2017.7.27.1
cffi==1.12.3
chardet==3.0.4
configparser==3.7.4
cryptography==2.7
Django==1.11.5
django-auth-ldap==1.2.15
django-constance==2.0.0
django-cors-headers==2.1.0
django-filter==2.0.0
django-guardian==1.4.9
django-picklefield==1.0.0
django-storages==1.7.1
django-suit==0.2.26
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
idna==2.6
Markdown==2.6.9
mysqlclient==1.3.12
paramiko==2.6.0
pyaml==19.4.1
pycparser==2.19
PyJWT==1.5.3
pyldap==2.4.37
PyNaCl==1.3.0
pytz==2018.7
PyYAML==5.1.1
requests==2.18.4
six==1.10.0
urllib3==1.22
websocket-client==0.44.0
psycopg2-binary